requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==5.2.9
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.2.1
django-unfold==0.69.0
djangorestframework==3.16.1
drf-yasg==1.21.11
environs==14.5.0
exceptiongroup==1.3.1
inflection==0.5.1
kombu==5.6.1
marshmallow==4.1.1
packaging==25.0
pip-chill==1.0.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
